Split generic behavior extends diagnostics
